requirements/base.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
alembic==1.12.1
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
gunicorn==21.2.0
python-dateutil==2.8.2
